ioht_data_collector/requirements.txt
httpx==0.26.0
openant==1.3.3
paho-mqtt==2.1.0
clickhouse-connect==0.8.15
adafruit-circuitpython-ble-cycling-speed-and-cadence==1.1.17
